Add a new readiness enum value, ReadinessClosed.
As discussed on #telepathy with oggis, this new enum value represents the readiness state of the channel when it has been closed. It replaces the closed() signal which could cause race conditions with queued signal/slot connections. This means ReadinessDead is now only for when the channel dies unexpectedly, rather than being closed by the class user in a controlled way.