codegen: support 'array-depth' attribute on types
In the Qt world, we need to generate a named type for every level of
nested list that makes sense. I've implemented this by adding an
array-depth attribute to <tp:mapping> etc. indicating the maximum
number of levels of nested lists that make sense for this type, so:

<tp:mapping name="MessagePart" array-name="MessagePartList"
  array-depth="2">

generates types MessagePart, MessagePartList = QList<MessagePart>,
and MessagePartListList = QList<MessagePartList>. If array-depth is
omitted, its value can be assumed to be 1 if array-name is present,
or 0 otherwise.

For the moment, the extra "List" suffix for arrays of nesting depth >= 2
is hard-coded.