tests/: consistently build generated moc files in _gen
Also, in every directory, generate moc files in such a way that we can
build them either from the .h (if there is one) or the .cpp (if not).

Rationale: if the QObject is entirely local to one test, we don't need
to bother making a header just for moc's benefit - we can just run moc
on the .cpp, which contains both declaration and definition. However,
for "library" code that already needs a header, like
tests/peregrine/lib.{h,cpp}, the declaration that moc needs is in the
header.