StatefulDBusProxy: find out the unique name
For a stateful interface it doesn't make sense to follow name-owner
changes, so we want to bind to the unique name.

Currently I match telepathy-glib's implementation and use a
pseudo-blocking call; since our call is only to the dbus-daemon this
doesn't introduce any message reordering, and it avoids having to
delay the start of the introspection process for subclasses until
the unique name has been discovered.