Channel: add protected methods to look for the self-handle in local pending, and add it
While using Contact objects is nice from a high-level-API point of view,
it's not so convenient if all we want to do is to accept an invitation or
VoIP call.

The high-level semantics of moving yourself from local-pending to members
depend on the channel type, so subclasses should introduce wrappers for
these methods with a friendlier name, where appropriate.