ContactManager: don't second-guess the connection manager
Although we have flags to indicate which actions are likely to work, their
interactions are complex, so we should only reject attempts to take action
if the channel doesn't exist at all. In all other cases we should attempt
the D-Bus call, even if we expect it to fail, so the connection manager
can return an appropriate error.