StreamedMediaChannel: Fixes/improvements.
Do not added MediaStream objects to StreamedMediaChannel until they are ready
(have a contact object).
Make MediaStream a ReadyObject and use this to retrieve the contact (reducing
code complexity).
Changed MediaStream to be a QSharedData in order to use it with
QExplitlySharedDataPointer.
Moved MediaStream signals to StreamedMediaChannel and use the
QExplitlySharedDataPointer<MediaStream> in the signals instead of MediaStream *.
Some other fixes.
Changed call example accordingly.