Clean up and simplify header search paths
Ensuring that $(top_srcdir) and $(top_builddir) are first on the -I path
is always a good idea - it means we always use this library's headers,
and not those of an installed copy of telepathy-qt4 that might be on the
-I path (due to sharing a directory with the installed telepathy-glib).

AM_CPPFLAGS is passed to every compilation, so we can use it for both C
and C++. It's also exactly what we want when running moc.

It also seems a good idea to make it explicit that we're linking QtCore.