Account: Removed manager method and use QDBusConnection/busName to create account.
Rationale:
Before this change, to create a Account, the user would pass the
AccountManagerPtr as a param.
This way, it's desirable that manager always return the proper manager used on
Account creation, but this was not true in all cases, as the account manager
could be destroyed while the account was still alive.

Example:
    AccountPtr acc;
    {
        AccountManagerPtr am = AccountManager::create();
        ...
        acc = Account::create(am, path);
    }
    // am is detroyed - nobody referencing it, Account HAD a weakref to it
    Q_ASSERT(acc->manager()); // fail

As the AccountManager is supposed to be a singleton, there is no need for this
method, and to avoid misbehaviour let's remove it.