shared-ptr: Changed SharedPtr(T*) to be a templated constructor.
Rationale (Jonathon Jongsma):
[explicit inline SharedPtr(T *d) : d(d) { if (d) { d->ref(); } }]
boost::shared_ptr uses a templated constructor for this case (e.g.  template
<class Y> SharedPtr(Y *d) ...) which allows it to somehow remember the actual
pointer type passed so that if it was a different type that is convertible to
T*, it will call delete with its original type and thus will not leak even if Y
does not have a virtual destructor). Granted, they probably use some black magic
to accomplish this, but it might be worth considering.