shared-ptr: Make SharedPtr(WeakPtr) constructor explicit.
Rationale (Jonathon Jongsma):
[inline SharedPtr(const WeakPtr<T> &o)]
It's probably best to make this explicit as well so that converting from a weak
pointer to a shared pointer is only done when it's intentional. You also might
consider something like the boost::weak_ptr::lock() API that allows you to
easily create a SharedPtr from a WeakPtr.