ClientRegistrar: Add multiple interfaces support on Client.
Removed AbstractClient, as there is no way to multiple inherit and use SharedPtr
at the same time, in favor of ClientObject.
To implement multiple interfaces clients will create a ClientObject and set it's
handler, approver, observer accordingly.
For now ClientObject only support handlers, but approver and observer support
will be added in the future when they exist.