MethodInvocationContext: Added class.
MethodInvocationContext is a class used to represent a context of a D-Bus
method call. It's main use is to handle async replies, where when the reply is
known, the methos MethodInvocationContext::setFinished/WithError should be
called, which will send the reply over D-Bus.

It's a template class to avoid replying with unknown/wrong types. It accepts all
types known by Qt MetaType System.