Fix compilation with Automake <= 1.11
TPQT4_V_MOC_ was defined in terms of $(AM_DEFAULT_VERBOSITY) which is
of course not defined with older automakes, so expanded to nothing, so
TPQT4_V_MOC_ was defined to be … itself. This patch effectively defaults
$(AM_DEFAULT_VERBOSITY) to 0 in older Automakes.

I tried to make all the definitions conditional on
$(AM_DEFAULT_VERBOSITY) being defined, but failed because this Makefile
is included by .am files, so Automake sees the 'ifdef' / 'endif' pair,
doesn't recognise 'ifdef' but does recognise 'endif' so gets upset.

Reviewed-by: André Moreira Magalhães <andre.magalhaes@collabora.co.uk>