Fix strict QtDBus demarshalling which was causing crashes when qdbus_casting from a SocketAddressIP*.
A specification for qdbus_cast<SocketAddressIPv4/v6> was implemented which checks for the incoming type signature:
if the incoming struct is (su), it uses a private helper structure which is then converted to the requested type.
It also prevents crashes in case the signature is not the expected one.

The nice thing is that no code modification are required: you just need to include the brand new types-internal.h.