Always use NULL instead of 0 in vararg lists
While 0 works fine in real C++ APIs which have, like, typing (:O), in
vararg lists putting 0 as an argument will just push a 32 bits zero
integer on amd64, although it should be a pointer argument of 64 bits.
This will a) mess up the following arguments b) if used as the last
argument, half of it will be zero and half of it uninitialized which is
bad for eg. g_object_new which may then go happily scanning for more
arguments since it's not really an all-zero pointer necessarily.

Detected by check-valgrind.