Fix TestAccountBasics race expecting avatarChanged
The avatarChanged signal might have already been caught when we were
making FeatureAvatar ready for the second time or when making the
SetAvatar call in the first place, so only start waiting for it if we
haven't got it already.

This fixes a test hang which happens quite often with higher CPU load
and check-valgrind (when the test itself runs slower, and the avatar may
change faster than the test is able to catch it).