Make Connection::contactManager() less of a death-trap
Previously the accessor used to return NULL if the Connection Core
wasn't ready. This is less of a problem for the case of the Connection
not being ready *yet* but very much one due to the fact that when the
Connection goes invalid it's not considered ready anymore.

When the conn goes invalid, application code (actually, in this case,
the Channel internals) should react by not making any more calls to the
Connection - when it gets the signal! However, if it's executing its own
code calling Connection *when* the Conn gets invalidated, it won't get
the signal in time to prevent doing so (it will only get the signal when
it returns to the mainloop). With the previous NULL return behavior,
applications would crash with a SIGSEGV in this case. Even if the signal
is caught in time preventing all further calls to avoid SIGSEGV will be
VERY hard, for any non-trivial application anyway.

This affected eg. the conference channel test, which disconnects the
connection and runs a mainloop until it emits invalidated. However, at
this point it still has Channel objects for the same connection, which
haven't all completed introspecting themselves. The Channel introspection
process will make calls to ContactManager to construct contacts for its
participants. If this happens in the same mainloop iteration as the
connection invalidation, but before the channels get the invalidated
signal, the channels' call to ContactManager will cause the application
to crash with a SIGSEGV.

This patch makes Connection::contactManager() always return a non-NULL
value, but will make ContactManager::contactsFor{Handles,Identifiers}()
and ContactManager::upgradeContacts() fail instantly with
TELEPATHY_ERROR_NOT_AVAILABLE and a descriptive debug message, if the
connection has been invalidated or never has been made ready.