Fix yet another race in the roster test
The previous fixes only really fixed the first contact in the request presence subscription test,
which does eventually accept our request. The second contact refuses our request, which, if it
happened before the test got to checking if it had asked for the subscription in the first place,
caused the test to fail (because it wasn't anymore in the "asked for subscription" state, but
"not subscribed").

This race only happened when running under valgrind, not without it even with some concurrent CPU
load.