Fix race in TestStreamedMediaChan::testOutgoingCallBusy
Tp-qt4 may have already noticed the remote is busy and invalidated the channel, so we should only
start waiting for the channel to become invalid unless it is still valid. Otherwise there's a
probablity (which is quite high under valgrind) that the test will hang, because it doesn't execute
fast enough to reach starting to wait for the invalidation before the channel has already been
invalidated.