Fix race in TestStreamedMediaChan::testOutgoingCallTerminate()
This required writing a fairly elaborate state machine, because the requestStreams pending
operation can finish anywhere between starting the call, the remote contact being added to RP, them
answering, and them terminating the call.

Should this be considered a bug, making applications write error-prone code? Probably not, as I
don't think the applications need to care about whether the group membership changes happen before
or after the call starting being signaled. They'll probably just use the membership change signals
to show a participant list, if anything. However, if they want to display a widget with the remote's
avatar, for example, when the PendingOp finishes, then indeed this should be considered a bug. (They
could do so anyway reliably though, by storing the contact they use to make the requestStreams call
in the first place, which is the most logical way to do it).

testOutgoingCall() itself is still racy, fixing it next.