Fix races in TestStreamedMediaChanCall::testOutgoingCall
The race was because the simulation connection service decides to start sending us video or stop
doing that based on the exact timing when we stop sending from our end after initiating the call.
The fix is to make sure we're neither sending nor receiving before starting to poke the stream back
and forth to ensure we don't eg. try to make it receive and wait for it do start doing so when it is
already in fact doing that.