Make Channel::groupSelfContact() always have some non-NULL contact if ready
This has the side effect of making emitting groupMembersChanged with people only in the removed set
work (again, I hope) even if the local user is not a member of the group. Actually... that bug has
probably made contact list groups totally suck, with them never reporting removed members! (We of
course usually aren't a member on the list of our contact groups, unless we consider extremely
lonely users the norm).