qt4-types-gen: improve broken spec error reporting.
This is partly just me being a bit OCD, but it makes the error output
way easier to read. Previously, if the generator hit (for instance) an
undefined type, it'd spit out:

      GEN    _gen/types-body.hpp
    Traceback (most recent call last):
      File "../../../tools/qt4-types-gen.py", line 490, in <module>
        Generator(dict(options))()
      File "../../../tools/qt4-types-gen.py", line 171, in __call__
        self.provide_all()
      File "../../../tools/qt4-types-gen.py", line 312, in provide_all
        assert not self.required_custom, 'These required types were not provided by the spec: ' + ', '.join(self.required_custom)
    AssertionError: These required types were not provided by the spec: CurrencyAmount

But now it produces:

      GEN    _gen/types.h
    Your spec is broken, dear developer! The following types were used, but not provided by the spec or by <tp:external-type/> declarations in all.xml:
        CurrencyAmount

A more informative error, and 100% less backtrace cruft.

Of course, it'd be better if we could report the locations where the
unknown types were used, but that seems to be non-trivial.