high-level-tubes: Instead of offering a QIODevice when accepting tubes, provide just the addresses.
Some applications need to use the bare address, because they use a
protocol library which doesn't communicate over QIODevices, but rather take
the address as a parameter and use internal implementation means to actually
establish a socket connection accordingly. This is because libraries
usually aim to be as generic as possible, while QIODevices require e.g. running
the Qt event loop.