Connection: Remove superfluous introspection status change guard
This was as far as I can see totally redundant with the ReadinessHelper mechanism to wait for the
current status to be introspected fully before moving to the new one. It might also have had a race
condition causing the assert hit crashes reported recently, although I can't spot or reproduce it.