Make the Channel main property extract fn more appropriately named
I originally planned to have one extraction function for each set of properties introduced in new
spec releases, queuing appropriate fallbacks if the properties introduced in that version weren't
found. Somebody didn't follow that plan, which would be useless to try and attempt now, so remove
the confusing references to it.