ContactList roster: Make roster groups actually work
The roster code failed to query the contact attributes telling their groups correctly, because
ContactManager::supportedFeatures() didn't know about the corresponding Contact feature.

The unit test for roster groups was interesting. It had the following paradigm:

- create group B
- copy all contacts which were in group A to group B
- verify that group A and B have the same contacts

This, however, passes even if the code fails to discover ANY contacts in group A (because an empty
group is equal to another empty group).