Import Dario Freddi's old low-level stream tubes example
I'll use this as a basis for testing and demonstrating the new tubes convenience API.
As the first thing I'll port it to use an Account, not ConnectionManager connection requesting
directly.

When the tube handling convenience API is done, I'll make it an exemplary user of that API:
requesting the tubes from the Account and handling them using StreamTubeServer/StreamTubeClient.