PendingContacts: Use the connection as the parent, not the contact manager
This is to avoid crashes I've seen occasionally in corner cases when developing tests, resulting
from the PendingContacts using the connection during some of the async contact build steps, but the
manager and thus the operation only having a weak pointer to it, and the conn being destroyed before
those steps. Making the connection the parent keeps it alive for all of the steps and because the
conn holds a strong ref to the manager, it's kept alive at least as long as before as well.