Deprecate external usage of PendingOperation::object()
The object() mechanism's sole purpose should be to keep the necessary parent object hierarchy alive,
and for the operation's internal code to have access to it.

It's not suitable for public use, because:

We don't guarantee which object is actually returned in any case, and even if we did, it would
require awkward type conversions AND checking their results to safely use the accessor for accessing
the object.