Call AbstractClient virtual base class constructor in our concrete Client impls
As AbstractClient is inherited virtually, nothing calls the ctor automatically. I noticed the
pointer to the AC base class is <invalid> as reported by gdb for all our concrete Client
implementations. This hasn't caused any problems so far however, because the base class is actually
empty, so the pointer is not used for anything. It will be used in the near future however when we
move the isRegistered stuff from Andre's cd-claim branch to the AbstracClient base class in an ABI
break.