Channel: Properly check if FeatureCore is ready.
When calling ReadyObject::isReady() without arguments the core feature passed to the ReadyObject
constructor is implicitly added to the list of features being checked.
Thus for channel subclasses when calling channelObject->isReady(), what is being checked
is whether the core feature of the subclass is ready, not Channel::FeatureCore, thus explicitly passing
Channel::FeatureCore when checking if this specific feature is ready is needed.