Account: Properly check if FeatureCore is ready.
When calling ReadyObject::isReady() without arguments the core feature passed to the ReadyObject
constructor is implicitly added to the list of features being checked.
Thus for account subclasses when calling accountObject->isReady(), what is being checked
is whether the core feature of the subclass is ready, not Account::FeatureCore, thus explicitly passing
Account::FeatureCore when checking if this specific feature is ready is needed.