Connection: Properly check if FeatureCore is ready.
When calling ReadyObject::isReady() without arguments the core feature passed to the ReadyObject
constructor is implicitly added to the list of features being checked.
Thus for connection subclasses when calling connectionObject->isReady(), what is being checked
is whether the core feature of the subclass is ready, not Connection::FeatureCore, thus explicitly passing
Connection::FeatureCore when checking if this specific feature is ready is needed.