Remove separation from client and service side high-level APIs.
To be more consitent with how client side is implemented, the service side high-level
classes are now moved to the TelepathyQt dir, removing the separate Service dir.
Library specific globals (TP_QT_EXPORT, etc) are now shared between tp-qt-service and tp-qt.
The plan is to merge both libraries into tp-qt in the future, when the service side bindings
are API/ABI stable.