Add classes for the callback system.
The callback system will be used initially by service side high-level APIs to expose methods
that may/should be overriden in implementations.
Ideally we'd use virtual methods for this, but as new methods can be added freely (when interfaces
changes), we would not be able to guarantee a stable API/ABI.

Other options were also analysed, such as virtual padding, virtual_hook as well as using Qt slots
override mechanism with QMetaObject calls, but in the end they all have their own issues and we
decided to go with callbacks.

Usage example:
  // expects a method with the following signature:
  // void method();
  Tp::Callback0<void> cb = memFun(myObj, &MyImpl::method); // or
  Tp::Callback0<void> cb = ptrFun(&method);

  // expects a method with the following signature:
  // char *index(const char *s, char c);
  Tp::Callback2<char*,const char*,char> cb = memFun(myObj, &MyImpl::index); // or
  Tp::Callback2<char*,const char*,char> cb = ptrFun(&index);