Properly build against Qt 5 alpha.
Qt 5 alpha release has some build issues that should be addressed in order to build tp-qt against
it:
- When running configure in Qt 5 alpha, the pkgconfig files are properly generated, but when make is
executed, the files are replaced with empty ones, so to build tp-qt one must either backup the pkgconfig
files generated during the configure phase before running make and replace them afterwards,
or re-run configure with the same parameters after building it. (reported and should be fixed when
beta is out)
- If Qt 5 alpha is built with -reduce-relocations enabled, one must add the keyword
"reduce_relocations" to qtbase/lib/pkgconfig/QtCore.pc in the qt_config section, so that tp-qt knows
when to use -fPIC/PIE accordingly. (reported and should be fixed when beta is out)