Move generated struct to TpDBus namespace
Only structs are moved because many of them are wrapped anyway and thus
the potential for collision is high (see rename of SimplePresence to
Presence).

Enums and flags are not moved because lots of code outside telepathy-qt
is using them directly and if there ever is a collision, the API would
become very awkward anyway with two types that are named the same in two
different namespaces doing different things.