BaseConnection::ensureChannel(): Refactored.
Channel matching extracted to a protected virtual method.
Previous ensureChannel() implementation doesn't allow to
override the default "ChannelType/TargetHandleType/TargetHandle
matching" behavior, which might be incorrect for some reasons.

This way, we can not perform the request values extraction outside of
the cycle, but it is a necessary sacrifice with low performance hit,
because ensureChannel() is rarely used method, extraction operation
is low-cost and usualy there is just a few channels to iterate.