doxygen-header: remove the 'Functions' link
doxygen no longer generates a single page for all functions;
instead, it generates one page per starting letter, so
we actually need one link for each letter of the alphabet.
Doxygen's own header normally generates a menu there, but
doing this is not very easy in a custom header and since the
functions list is not very useful anyway, I'm removing it for now.