Fix wrong version number matching
As the CMake MATCH takes regex parameter, sub-string could unexpectedtly
be matched. QT verseion 5.12.4, for instance, will be matched with not just
for QT5 but also for QT4 as well due to the sub-string '4' which is unexpected.
This patch resolves the issue by constraining version number testing for
Major version numbers only.

Signed-off-by: Brandon Hong <brandon.hong@intel.com>